// src/app/services/openapi-client/model/userDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDto { 
    id?: number;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    readonly fullName?: string | null;
    phone?: string | null;
    role?: string | null;
    status?: string | null;
    emailVerified?: boolean;
    createdAt?: string | null;
    lastLoginAt?: string | null;
}